+++ src/east/__main__.py
@@ -3,7 +3,16 @@ import rich_click
 
 from .east_context import EastContext, east_group_settings
 from .system_commands import sys_setup, util
-from .workspace_commands import build, bypass, clean, debug, flash, release, update
+from .workspace_commands import (
+    build,
+    bypass,
+    clean,
+    debug,
+    flash,
+    release,
+    twister,
+    update,
+)
 
 rich_click.rich_click.MAX_WIDTH = 80
 rich_click.rich_click.USE_RICH_MARKUP = True
@@ -21,6 +30,7 @@ rich_click.rich_click.COMMAND_GROUPS = {
                 "update",
                 "bypass",
                 "release",
+                "twister",
             ],
         },
         {
@@ -77,6 +87,7 @@ cli.add_command(sys_setup)
 cli.add_command(util)
 cli.add_command(release)
 cli.add_command(debug)
+cli.add_command(twister)
 
 
 def main():